Propagate errors from process::{get_process_id, get_module_base}
This updates the signatures of these two functions to return a Rust std::io::Error type, deriving its contents from the `Error::last_os_error()` function.

Another reasonable approach might be to return a custom instance of std::io::Error, or create an error type for the process module, but this felt like a thin enough change for now.